restruct token_status
