+++ src/oid4vci/handlers/credential_offer.rs
@@ -18,8 +18,8 @@ use anyhow::Context as _;
 
 use crate::oid4vci::error::invalid;
 use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
-use crate::oid4vci::provider::{Provider, StateStore};
 use crate::oid4vci::issuer::{CredentialOffer, CredentialOfferRequest, CredentialOfferResponse};
+use crate::oid4vci::provider::{Provider, StateStore};
 
 /// Endpoint for the Wallet to request the Issuer's Credential Offer when
 /// engaged in a cross-device flow.
