+++ src/oid4vci/handlers/issuer.rs
@@ -26,8 +26,8 @@
 use anyhow::Context as _;
 
 use crate::oid4vci::handlers::{Body, Error, Handler, Headers, Request, Response, Result};
-use crate::oid4vci::provider::{Metadata, Provider};
 use crate::oid4vci::issuer::{IssuerRequest, IssuerResponse, MetadataHeaders};
+use crate::oid4vci::provider::{Metadata, Provider};
 
 /// Metadata request handler.
 ///
