+++ src/oid4vci/handlers/nonce.rs
@@ -12,9 +12,9 @@ use chrono::Utc;
 
 use crate::generate;
 use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
+use crate::oid4vci::issuer::{NonceRequest, NonceResponse};
 use crate::oid4vci::provider::{Provider, StateStore};
 use crate::oid4vci::state::Expire;
-use crate::oid4vci::issuer::{NonceRequest, NonceResponse};
 use crate::state::State;
 
 /// Nonce request handler.
