+++ src/oid4vci/handlers/par.rs
@@ -11,8 +11,8 @@ use chrono::{Duration, Utc};
 
 use crate::generate;
 use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result, authorize};
-use crate::oid4vci::provider::{Metadata, Provider, StateStore};
 use crate::oid4vci::issuer::{PushedAuthorizationRequest, PushedAuthorizationResponse};
+use crate::oid4vci::provider::{Metadata, Provider, StateStore};
 use crate::state::State;
 
 /// Endpoint for the Wallet to push an Authorization Request when using Pushed
