+++ src/oid4vci/handlers/register.rs
@@ -4,9 +4,9 @@ use anyhow::Context as _;
 
 use crate::oid4vci::error::server;
 use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
+use crate::oid4vci::issuer::{RegistrationHeaders, RegistrationRequest, RegistrationResponse};
 use crate::oid4vci::provider::{Provider, StateStore};
 use crate::oid4vci::state::Token;
-use crate::oid4vci::issuer::{RegistrationHeaders, RegistrationRequest, RegistrationResponse};
 
 /// Registration request handler.
 ///
