+++ src/oid4vci/handlers/server.rs
@@ -19,8 +19,8 @@
 use anyhow::Context as _;
 
 use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
-use crate::oid4vci::provider::{Metadata, Provider};
 use crate::oid4vci::issuer::{ServerRequest, ServerResponse};
+use crate::oid4vci::provider::{Metadata, Provider};
 
 /// OAuth server metadata request handler.
 ///
