+++ src/oid4vci/handlers/token.rs
@@ -23,13 +23,13 @@ use crate::generate;
 use crate::oauth::GrantType;
 use crate::oid4vci::error::{invalid, server};
 use crate::oid4vci::handlers::{Body, Error, Handler, Request, Response, Result};
-use crate::oid4vci::pkce;
-use crate::oid4vci::provider::{Metadata, Provider, StateStore};
-use crate::oid4vci::state::{Authorized, Expire, Offered, Token};
 use crate::oid4vci::issuer::{
     AuthorizationCredential, AuthorizationDetail, AuthorizedDetail, Issuer, TokenGrantType,
     TokenRequest, TokenResponse, TokenType,
 };
+use crate::oid4vci::pkce;
+use crate::oid4vci::provider::{Metadata, Provider, StateStore};
+use crate::oid4vci::state::{Authorized, Expire, Offered, Token};
 use crate::state::State;
 
 /// Token request handler.
