+++ src/oid4vp/wallet.rs
@@ -1,9 +1,9 @@
-//! # Wallet 
+//! # Wallet
 
+mod metadata;
 mod response;
 mod vp_token;
-mod metadata;
 
+pub use metadata::*;
 pub use response::*;
 pub use vp_token::*;
-pub use metadata::*;
