+++ src/oid4vp/wallet/vp_token.rs
@@ -1,7 +1,5 @@
 //! #! Querying credentials
 
-
-
 use std::collections::HashMap;
 
 use anyhow::Result;
